styles: Move new flex-column-reverse style off of "context".
This doesn't (and shouldn't) depend on the theme, so there isn't a
need for it to be passed through context.

It's also easier to understand the places it's used if it's right
there.  This isn't a case where there's any meaningful abstraction
that the name can express; at best the name is just a restatement of
the actual source code of the style.