styles: Consolidate settingsIcon style.
There was a misleading global `icon` style, which in reality we only
used in a couple of the various places that we use icons and put
styles on them.  Conveniently, we used it in the exact same way in
each of them, combining it in each case with the static
`styles.settingsIcon`, which conversely was only ever used in those
places... so fold all its static properties into `settingsIcon`, and
then use the pattern of the preceding few commits to handle the one
theme-dependent property.