account: Replace empty-string fallback with assert, in getIdentity.
This selector actually has zero callers right now -- so we're free to
do with its semantics whatever we like!

(We introduced this selector and the `Identity` type recently, and
haven't yet made the effort to adopt it widely.  We'll add a new
caller shortly, as we move callers off the empty-string semantics of
getAuth.)