account: Make `getAuth` return only real, authentication-bearing values.
I read through all the call sites of `getAuth` (that remain after the
changes in the last half-dozen commits), and all of them other than
these two make an assumption that the returned Auth object has a real,
valid API key -- either they're in part of UI that displays logged-in
information, or they promptly use the return value to make an
authenticated request to the server, or both.

These two don't really want an "auth" object so much as a realm and
optional email, because they actually assume we don't have a valid
API key -- they're both part of authentication flows.  But they want
an `Auth` object to pass to the API bindings, so unpicking that will
be a bit more work; just make it a "partial auth" that is still of
type `Auth`.