account: Cut remaining direct use of Account type in application code.
This UI has no need to pass around the actual API key for each
account; it only needs the boolean of whether we have an API key or
not.  Make a quick selector to provide that information, and pass that
result around instead.

This will also simplify a coming refactor that adds data to Account,
by making fewer places that need to update with it.