notif: Simplify by pushing maybeSend logic up the call chain.
This happens exactly the same way in the iOS and Android cases, so
take care of it once in the common code.  As a bonus, this simplifies
the signatures quite a bit.

The sharp-eyed reader will spot a couple of changes here that locally
look like possible behavior changes, both on the Android side:

 * We call `refreshToken` in a broader range of conditions.
   That's fine; it's harmless to overdo.

 * We call `refreshToken` after `maybeSendPushToken`, rather than
   before.  Its behavior is already async, so this won't cause any
   behavior that wasn't already possible.  And its only possible
   visible effect is to cause `handleDeviceToken` to get called, and
   therefore `sendAllPushToken`; in that case, it doesn't matter
   whether `maybeSendPushToken` was called before, after, or at all.