i18n: Move message files out of src/, into static/.
These are a common source of noise when grepping the codebase,
especially for terms like "Stream" or "Message" -- anything that
matches even one message string piles up a ton of hits as it appears
in each of these files.

The worst part is that there hasn't been an easy way to filter this
noise out when doing a grep, because these files appear right in the
middle of our main source code.

Fix that by moving them into static/, so that `git grep Stream src/`
etc. find only the matches that are actually in our source code.