language picker: Let first tap make a selection and dismiss keyboard.
When the search box was focused and the keyboard visible, tapping a
language would just unfocus the input and hide the keyboard, without
actually selecting the new language; you'd have to tap a second time
to make a selection.

This is the default behavior of `ScrollView`, found inside the
`FlatList` we use to show the languages.

Instead, set `keyboardShouldPersistTaps` to `always`.  This disables
`ScrollView`'s handling of the tap, letting our list item receive it.
  https://facebook.github.io/react-native/docs/scrollview#keyboardshouldpersisttaps

(The search box still gets unfocused and the keyboard hidden, even
though `ScrollView` is no longer doing anything to cause that.  This
seems to be triggered by the settings change... ah, and it's caused
by the hack in `TranslationProvider` described in a comment there,
where we change `key` when the locale changes.)

[greg: Wrote commit message.]