persist-migrate: Don't run old migrations on first install.
It turns out that `redux-persist-migrate` as taken from upstream wants
to run all the migrations from scratch when `payload` in the REHYDRATE
action is missing or empty -- which is what happens on first install.
That would make each migration permanently a critical part of our
code, on the path for every new user; and one that's rarely run in
normal use, making bugs likely to slip through.

In fact, there's no need for any migrations, and they wouldn't do much
anyway, because the rehydrated state (`action.payload` aka `incomingState`)
is an empty object.  The state in the Redux store has already been
initialized by the reducers to the right shape, from the `initialState`
object in each reducer, when the store was first created.

Instead, skip the migrations.  Just set the version to the current
one, so that future upgrades will run any new migrations they include.

This fixes an issue introduced in c2ed78e7e, where migration 2 tries
to update the shape of `state.realm.pushToken`, and on a fresh install
would crash because there is no such property.

[greg: expanded commit message]