reactions: Use user ID instead of email.
We've been identifying the user that posted a reaction by their
email, rather than their user ID.  (So far, we use this
information only to tell if one of the users with a given
reaction was this user themself.)  As with the app in general,
we should move to using the user ID.

In seeking to do so, we learned that the server is actually
inconsistent in how it describes the user on a reaction!  The
reactions in a /messages request have a `user.user_id`, while
those in a reaction event have a `user.id`.  We recorded that
fact in 2173b367a.

Really we don't want either of those structures, though: we just
want the user ID, and we'll get the other details from our
normalized other data structures as needed.  So, rather than deal
on the inside of the app with these two different "Reaction"
types, we define "Reaction" as the type we actually want, and
convert both variations to that one at the edge.

Part of #2252.

[greg: moved all type complications out of modelTypes, the
 interface; added some comments; rewrote commit message.]