settings: Bring initialNumToRender back to reasonable value in LanguagePicker
This effectively-infinite value for initialNumToRender was added as
a hack for an issue described in d2bf35ff3 / #1350, several years
and many RN versions ago.  Its effect is to totally disable the
normal benefits of FlatList.

Whether or not this was an appropriate hack then, there doesn't
seem to be a need for it now -- the issue doesn't reproduce if we
bring the prop back to a reasonable value.

The documentation recommends setting this prop like so:

  This should be enough to fill the screen but not much more.

  https://reactnative.dev/docs/flatlist#initialnumtorender

On a moderately-tall test device, I (Greg) get 11 items on the
screen, and the default for this prop is 10.  So we do actually
want somewhat more than the default; make it 15.

[greg: set to 15 instead of default; expand commit message]