actions: Fix remaining `any` types in `EventAction` union.
To avoid making this task bigger than it needs to be, we defer saying
anything about the *other* properties of these actions, other than the
one critical property `type`.

(Except in two cases, where we don't even have any code to consume
actions of these types; for those, conservatively say there *are* no
other properties.)

With this fix, Flow now knows enough about `EventAction` to do a
reasonable job of type-checking `eventToAction`... where it finds that
in a bunch of cases we're returning objects that don't fit in the type
we've defined.  For this commit, just make a fixme; we'll fix it
shortly.