events: Use type inference in eventToAction.
This is a lot like the recent "redux: Use type inference in reducers",
but for `eventToAction.js`, which has a very similar structure to our
reducer modules.

Like with the reducers, cutting these explicit type annotations makes
this code a lot shorter and simpler to read, which I'm happy about.

And like with the reducers, a crucial ingredient in making this work
is that Flow now understands this code, following our recent fixes to
action types:

  ea2b119aa actions: Make EventAction complete.
  e6868d833 actions: Clarify to Flow that these constants are type-tags.
  ... and some neighboring commits.