events: Inline all the tiny eventToActions handlers.
This is much like the recent commits

  redux: Inline lots of tiny handlers directly into reducers.
  redux: Inline some more tiny handlers into reducers.

but for eventToActions, which has a very similar structure to our
reducer modules.  It follows on the simplifications in the parent
commit to make this code much simpler to read.

As with the reducers, I'm especially happy about the reduced
indirection; and importantly, no type coverage is lost.  Before
and after this change, there are 10 expressions in this file
Flow identifies as not covered, all of them due to taking
properties from `event` which is of type `Object`.