chat screen: Fix broken layout on iOS.
Fixes #3273.

A previous commit fixing a weird issue on Android inadvertently
broke the iOS layout:
f142452b4 chat screen: Work around RN bug making nav bar fail to render.

The ZulipStatusBar component is a difference between Android and iOS
as it has actual height on iOS but not on Android, we can never paint
in the status bar area of Android, we can change the color only.

This commit keeps the fix from the previous commit but simplifies it:
* removed the extra View, no need for the extra wrapper
* simplified the style, no need for the extra `flex`
* moved `ZulipStatusBar` to bottom because of
  `flexDirection: 'column-reverse'