compose box: Cut a redundant `color` style attribute.
Looking at this `composeTextInput` style next to `topicInput`, and
right above the one place where each of them are used, a question that
immediately comes to mind is: why does one of these need `color` set
and the other doesn't?  Do we really show them with different text
colors?... no, empirically we don't.

In fact, the answer is that neither needs to set this attribute: our
`Input` component already takes care of it.

So we get to cut it out.