AutocompleteViewWrapper: Simplify by inlining this component.
This component has just one call site, and does just one thing: it
takes a whole bunch of props, and passes some of them to one
subcomponent and others to another.

Something like this could be useful if it were providing some kind of
meaningful abstraction; a good sign would be if it were possible to
write a sentence or two of jsdoc that clearly identified what one of
these components *means*.  But this one doesn't.  Among other signs of
that situation, not a single one of its props gets used more than once.

So, simplify by just inlining it.  The call site in `render` gets
longer by three lines -- namely `/>`, `/>`, and `</View>` -- and
rather clearer to read and understand, because there's one fewer layer
of indirection to trace through.