compose box: Narrow the scope of the debounce-focus hack.
We pay attention to whether the message *or* topic input is focused,
and we have a bit of a hack to deal with the fact that when the user
moves focus from one to the other the `blur` event comes before
`focus`: on blur, we wait 200ms before doing anything about it, to
give time for a `focus` to show up on the other input.

But currently this hack has some collateral damage:

 * Because autocomplete popups also consult `is*Focused` to decide
   whether they should be visible, the hack causes them to lag when
   the user moves focus to the message list, or between inputs.

 * Because this logic is a little subtle and is not at all made clear
   in the code, we've wound up with changes to `isMenuExpanded` being
   delayed by the same 200ms.  This causes the closing of the menu to
   lag when the user moves focus to the message list.

Fix these by letting `isMessageFocused` and `isTopicFocused` promptly
reflect the truth of whether the message and topic inputs,
respectively, really are focused.  Make a separate state property for
the debounced "or" of them, and limit the delay to just updating that
one property.