android: Properly enable Chrome debugging of WebViews.
This follows the instructions of Chrome upstream:
  https://developers.google.com/web/tools/chrome-devtools/remote-debugging/webviews

There's a bit of code in RN to do this, in ReactWebViewManager's
`createViewInstance` method... but

 (a) that is clearly the wrong place for it, because this setting is
     global to the application, not per-WebView -- it probably works,
     but suggests confusion about how this works;

 (b) it only applies when ReactBuildConfig.DEBUG is true -- and that
     turns out to mean *React Native itself* is from a debug build.
     (Which it normally isn't, even in debug builds of an app.)
     This is the RN bug facebook/react-native#11415 .

If we wanted to enable this only in debug builds of the app, the
upstream doc linked above offers a recipe for it.

For Zulip, we have no desire to hide the code from the user; and
WebView debugging is already only available when the device itself has
remote debugging enabled, so there's no privacy or security reason to
disable it here.  Just unconditionally enable.

---

I confess some confusion because I remember this *did* use to work for
me, as recently as a few weeks ago.  Not sure what triggered the
change.  In any case, every cached version of the RN jar I have lying
around, going back to 0.55.4, has this `DEBUG` value set to `false`:

  for j in  ~/.gradle/caches/*/*/react-native-*/*/jars/classes.jar; do
    echo; echo "$j"
    javap -constants -cp "$j" com.facebook.react.BuildConfig \
      | grep DEBUG
  done

(It turns out `ReactBuildConfig.DEBUG` is basically just an alias of
`com.facebook.react.BuildConfig.DEBUG`.)