notif: Better abstract the listen and unlisten operations.
The wix library gives one API in the iOS case, and a quite different
-- and much less reasonable -- one in the Android case.  The main
point of this change is to cover those with a single, reasonable,
abstraction.

This should be a pure refactor in the iOS case.

In the Android case, this ensures we don't end up with two of these
listeners for `notificationOpened`.  That isn't likely in practice,
because this is made by AppEventHandlers which is far to the outside
of our React tree... but better to make it right.

Also add some jsdoc -- mainly to clarify which of these methods are
intended as private helpers and which are the external interface of
this class, as we've added more private helpers.