notif: Further cut down the old getToken* logic.
It's not clear this failure listener is doing us a lot of good; but in
any case it should match the success case in where and when we listen.

Then the rest is small enough that we may as well inline them back
into the if/else they're called from.