user status: Add Redux action creators to make API calls for update.
Like many other redux actions that call API functions, currently the
main goal of these functions is  to extract the 'auth' value from the
state and pass it to the API call.

We use two functions instead of one universal that can pass both
values at once, for more clear, simple and robust API and because
just like the web app we do not update the `away` and `status_text`
values together.