tools/test: Drop the spellchecker by default.
I rarely find this useful, if ever; and @borisyankov mentioned a few
weeks ago that he didn't find it very helpful either.  On my desktop
it adds 500ms or so to `tools/test` in the best case, up to about 4s
if running on the whole tree.  Just skip it by default.

It's still easy to get at for anyone who wants to use it; just run

  tools/test spell

Or if you want to run it along with the other suites, run

  tools/test flow lint jest prettier spell

If that's a popular thing to do, it'd be easy to add an alias "suite"
with a name like `all`, so that e.g. `tools/test all` does that.