notif: Drop the use of the `config.startup` mechanism.
Instead, just rely on `doNarrow` to take us to the desired narrow.
Since 42adfb348 "store: Buffer thunk actions", this action (and for
that matter the `handleNotification(...)` action which will dispatch
it in the first place) is buffered until after rehydrate, and so the
ad-hoc buffering we had in the `rehydrate` handler of `navReducers`
is no longer necessary.

In fact, my commit 9d5c0a73a yesterday already caused this
`config.startup` mechanism to have no effect!  The line where we
set it got moved inside a thunk action (along with the reference to
`usersById` which was the main target of that change), so that it
only happens after rehydrate is complete... at which point the code
that would act on it, in that `rehydrate` handler, has already run.
And that's fine, for the reason above (and empirically was fine in
my testing.)

This change overlaps in part with the work sent by @borisyankov
in #3282.