android notif: Depend explicitly on GCM SDK in Gradle config.
We get away without this because both react-native-device-info
and the wix react-native-notifications pull it in; but we now refer
to this SDK's classes directly, so we should declare the dependency
directly too.

The form of this line is taken from:
  https://developers.google.com/cloud-messaging/android/client

We use 16.0.0 because that's the latest.  It's what we've actually
been using all along, thanks to react-native-device-info just saying
the version should be `+`.