caught-up: Convert NULL_CAUGHTUP to an explicit DEFAULT_CAUGHTUP.
This is an example of a "null" object where all our uses reflect a
common, meaningful semantics.

When a given narrow lacks an entry in the caught-up state, it doesn't
mean "we don't know", as it does for e.g. a user missing in our list
of users -- it means "we're not caught up on that narrow", because we
know nothing about the narrow, and being caught up would mean we know
we have the earliest and/or latest messages as of our current snapshot
from the event queue.  And "we're not caught up" is exactly what this
value says.

So, document that explicitly, and name the object accordingly.