presence: Add a discrete 'unavailable' presence status.
Previously, if the user was set to `away: true` we 'masked' their
presence to be indistinguishable from 'offline'.

The requirements were changed and to match what we do on the web,
introduce a new presence status 'unavailable' when user is away.

The change to 'PresenceStatusIndicator' is not visual but to
preserve the rendering as-is for now.

[greg: updated UserStatus jsdoc]