streamsActions: Fix vague and erroneous types; mark strict-local.
We've been passing a string where the signature we'd written says
only a boolean is permitted.  The only reason it'd been possible
for that error to survive is that we'd written the vague type
`Object` here.

In fact, the truth is rather peculiar, because the code at the call
site of this function is idiosyncratic.  Write the truth.