narrow: Assert users not missing in getRecipientsInGroupNarrow.
The existing code was ill-typed -- if a user was unknown, it would
substitute `[]` in place of a User object, as an element of the
returned array.  Naturally, both of the places where we use this
selector would promptly blow up if that happened.

Moreover, both look like places embedded inside UI that already
assumes that this PM conversation is well formed and consists of
valid known users; in which case, it's impossible for us to get
to this condition in any case.

Instead of blowing up in an uncontrolled fashion with a type error,
give at least a sane explicit exception in this case, if it can
in fact ever happen.  And return a proper type.