utils/tests: Remove crufty test-helper file.
The last call site was removed in 11833a665 a little short of a year
ago, with the explanation "delete StylesProvider-test.js as failing
and not useful".

---

This concludes a series of 36 commits that added `@flow strict-local`
to the majority of our remaining files that lacked it, fixing a
number of vague or erroneous types along the way.

Before this series (at ac4ea5bdf), 58 files lacked strict-local:

  $ tools/tsort-modules todo | grep -cv strict-local
  58

Now only 16 of them do:

  $ tools/tsort-modules todo | grep -cv strict-local
  16