Icons: Drop confusingly-named default export; give it a type.
We've had a bunch of imports of this module that take the default
export but then give it a name different from the module's name.
That's confusing; instead, just give that export the name we want
to use for it.

Also give that export a type, building on the recent work in 6b39648ab
where we first gave real types to this module.

And fix a type error that that exposes in a call site!  If we'd ever
actually failed to pass `name` to `NavButton`, nothing good would have
come of that.