avatar: Take a `size` parameter for all URL-generating functions.
We have been using `80` as a default size for avatars. After some
experimenting it still seems a decent default that we can't improve
upon without writing some more code to be more precise in all cases.

Size is important, so add this to all functions. This can be optional
on `getAvatarFromUser` and `getAvatarFromMessage` as they use
`getAvatarUrl` which has that set to a reasonable default of 80.