narrow: Propagate and comment the peculiar type of getStreamInNarrow.
Because of #3367 (an issue in the `flow-typed` libdef for
`react-redux`), Flow is unable to catch where our components
make too-strong assumptions about what this selector returns.
Fortunately the runtime code makes only the assumptions that are
actually justified; we've just written down types that assume
too much.  Fix those.

Also the truth about what this selector returns, as reflected now
(since the parent commit) in its type, is over-complicated; the logic
should be simplified.  Make a small TODO for that.