getGravatarFromEmail: Remove default value for `size`
No function relies on the default value so remove it.
Only `getAvatarUrl` now is a direct caller and it is guaranteed to
pass a value.

That also removes duplication as we might want to change the `80`
value to another default (we will, stay tuned) - this will allow us
to change it only in one place.

Also removed one of the tests that was testing the difference between
a default `size` value and one passed explicitly, now made redundant.