Touchable: Adjust `children` prop's type to match View.
Very similar to 2a42e97d2 which did this for `Screen`.

A call site I added for `Touchable` caused Flow to start complaining
that it couldn't figure out what type to treat its children as.  The
error looked like this (edited for brevity and formatting):

  Could not decide which case to select. Since case 1 [1] may work but
  if it doesn't case 2 [2] looks promising too. To fix add a type
  annotation to [... one of the child elements, or the other, or the
  other.]

    /tmp/flow/flowlib_41860ea/react.js
  [1] 254│   declare export type ChildrenArray<+T> = \
                 $ReadOnlyArray<ChildrenArray<T>> | T;

    src/common/Touchable.js
  [2]  16│   children: ChildrenArray<*>,

Well, what `Touchable` does with `children` is just to pass it
straight on to `View`.  So I consulted the type annotation on
`View`... and it says `React.Node`.  (For which we have an alias
`React$Node`.)  So, match that.