api types: Fix the name of CrossRealmBot, from RealmBot.
The old name describes pretty much the opposite of what this type is
supposed to be: it's for *cross*-realm bots.  That means bots that are
in *all* (or several, I guess) realms -- the one thing it definitely
wouldn't include is a bot that is specific to this realm.

This confused me for quite a while when reading code that mentioned
this type.  Make it more truthful in hopes of being less confusing.